Use @KafkaListener istead of manual configuration in ACI test

+++ src/test/kotlin/no/nav/veilarbaktivitet/aktivitetskort/graphql/AktivitetskortControllerTest.kt
@@ -194,7 +194,7 @@ class AktivitetskortControllerTest: SpringBootTestBase() {
                 }
             }
         """.trimIndent().replace("\n", "")
-        val result = aktivitetTestService.queryHistorikkRaw(mockBruker, mockBruker, query, aktivitet.id.toLong())
+        val result = aktivitetTestService.queryHistorikkRaw(mockBruker, mockBruker, query, aktivitet.id)
         val fraDatoString = JsonMapper.defaultObjectMapper().readTree(result)["data"]["aktivitet"]["fraDato"].asText()
         assertThat(fraDatoString).isEqualTo(fraDatoIso)
     }
